developing with lue

+++ model/luepcr/operations.py
@@ -141,7 +141,7 @@ def lddrepair(arg):
 
 def numpy2pcr(arg1, arg2, arg3):
     print("Note 'pcraster.numpy2pcr' function is PARTIALLY implemented on LUE yet.")
-    return lfr.from_numpy(np.transpose(arg2), partition_shape, arg3)
+    return lfr.from_numpy(np.transpose(np.float32(arg2)), partition_shape, arg3)
 
 
 def Scalar():
